extract legacy static api and result modules
