+++ misda/__init__.py
@@ -49,6 +49,14 @@ from ._pareto import (
     evaluate_pareto_raw,
 )
 from ._validation import _validate_input_matrix
+from .result import MISCandidate, MISDAResult
+from .api import (
+    report_significant_correlations,
+    misda_significance_from_corr,
+    misda_significance,
+    _analyze_static_fast,
+    _analyze_static,
+)
 
 __version__ = "0.4.2"
 
@@ -401,199 +409,10 @@ def plot_custom_misda_graph(
 # Ranking (reexported from misda._ranking)
 # Core ISDA
 
-def report_significant_correlations(R, z_stat, z_crit, max_pairs=50, label_prefix="f"):
-    """
-    Generates a string report of significant correlations found in the data.
-
-    Args:
-        R (np.ndarray): The correlation matrix.
-        z_stat (np.ndarray): The Fisher z-transformed correlation statistics.
-        z_crit (float): The critical z-value for significance.
-        max_pairs (int): Maximum number of significant pairs to report for each type (positive/negative).
-        label_prefix (str): Prefix for feature labels (e.g., "f" for f1, f2).
-
-    Returns:
-        str: A formatted string report of significant correlations.
-    """
-    M = R.shape[0]
-    pos_corr = []
-    neg_corr = []
-
-    for i in range(M):
-        for j in range(i + 1, M):
-            if abs(z_stat[i, j]) > z_crit:
-                rij = R[i, j]
-                if rij > 0:
-                    pos_corr.append((i, j, rij))
-                elif rij < 0:
-                    neg_corr.append((i, j, rij))
-    
-    out = []
-    out.append("\n--- SIGNIFICANT CORRELATIONS (Fisher z, two-tailed) ---")
-
-    if pos_corr:
-        out.append("\nSignificant POSITIVE correlation:")
-        for i, j, r in pos_corr[:max_pairs]:
-            out.append(f"  {label_prefix}{i+1} – {label_prefix}{j+1}:  ρ = {r:.4f}")
-        if len(pos_corr) > max_pairs:
-            out.append(f"  ... ({len(pos_corr) - max_pairs} pairs omitted)")
-    else:
-        out.append("\nSignificant POSITIVE correlation: none")
-
-    if neg_corr:
-        out.append("\nSignificant NEGATIVE correlation:")
-        for i, j, r in neg_corr[:max_pairs]:
-            out.append(f"  {label_prefix}{i+1} – {label_prefix}{j+1}:  ρ = {r:.4f}")
-        if len(neg_corr) > max_pairs:
-            out.append(f"  ... ({len(neg_corr) - max_pairs} pairs omitted)")
-    else:
-        out.append("\nSignificant NEGATIVE correlation: none")
-        
-    return "\n".join(out)
 
 
-def misda_significance_from_corr(corr, N, M, alpha=0.05, labels=None, ensure_coverage=True, min_coverage=None):
-    """
-    Executes MISDA logic given a precomputed correlation matrix.
-    """
-    if labels is None:
-        labels = [f"f{i+1}" for i in range(M)]
-
-    corr = np.clip(corr, -0.999999, 0.999999)
-
-    z = 0.5 * np.log((1 + corr) / (1 - corr))
-    sigma_z = 1 / np.sqrt(N - 3)
-    z_stat = z / sigma_z
-
-    z_crit = stats.norm.isf(alpha / 2)
-    z_threshold = z_crit * sigma_z
-    r_crit = np.tanh(z_threshold)
-    if r_crit >= 0.99999:
-        r_crit = 0.99999
-    
-    corr_report = report_significant_correlations(corr, z_stat, z_crit, label_prefix="f")
-
-    signif = (_correlation_strength(z_stat) > z_crit)
-    adjacency = signif.astype(int)
-    np.fill_diagonal(adjacency, 0)
-
-    visited = [False] * M
-    components = []
-
-    def dfs(start):
-        stack = [start]
-        comp = []
-        while stack:
-            i = stack.pop()
-            if not visited[i]:
-                visited[i] = True
-                comp.append(i)
-                neighbors = np.where(adjacency[i] == 1)[0]
-                for j in neighbors:
-                    if not visited[j]:
-                        stack.append(j)
-        return sorted(comp)
-
-    for i in range(M):
-        if not visited[i]:
-            components.append(dfs(i))
-
-    components_labels = [[labels[i] for i in comp] for comp in components]
-
-    mis_sets = find_maximal_independent_sets(adjacency)
-    
-    if ensure_coverage:
-        repaired_sets = []
-        coverage_threshold = min_coverage if min_coverage is not None else r_crit
-        for ms in mis_sets:
-            repaired = repair_mis_coverage(corr, ms, min_coverage=coverage_threshold)
-            if repaired not in repaired_sets:
-                repaired_sets.append(repaired)
-        mis_sets = repaired_sets
-
-    mis_sets_labels = [[labels[i] for i in mis] for mis in mis_sets]
-
-    mis_metrics = compute_mis_metrics(mis_sets, adjacency, labels)
-    mis_sorted = sort_mis_metrics(mis_metrics)
-
-    def rank_key(m):
-        return (
-            m["neighborhood"],
-            m["span"],
-            round(m["avg_external_degree"], 4),
-            m["size"],
-        )
-
-    mis_ranked = []
-    rank_groups = {}
-    current_rank = 0
-    prev_key = None
-
-    for m in mis_sorted:
-        k = rank_key(m)
-        if prev_key is None or k != prev_key:
-            current_rank += 1
-            prev_key = k
-        m_with_rank = dict(m)
-        m_with_rank["rank"] = current_rank
-        mis_ranked.append(m_with_rank)
-        rank_groups.setdefault(current_rank, []).append(m_with_rank)
-
-    best_mis_rank1 = rank_groups.get(1, [None])[0]
-    best_mis_rank2 = rank_groups.get(2, [None])[0] if 2 in rank_groups else None
-
-    unique_metric_values = {
-        "neighborhood": sorted({m["neighborhood"] for m in mis_metrics}),
-        "neighborhood_ratio": sorted({m["neighborhood_ratio"] for m in mis_metrics}),
-        "span": sorted({m["span"] for m in mis_metrics}),
-        "avg_external_degree": sorted({m["avg_external_degree"] for m in mis_metrics}),
-        "avg_internal_degree": sorted({m["avg_internal_degree"] for m in mis_metrics}),
-    }
-
-    min_compactness, component_metrics, homogeneity_stats = calculate_component_compactness(corr, components)
-
-    return {
-        "corr": corr,
-        "adjacency": adjacency,
-        "components": components,
-        "components_labels": components_labels,
-        "mis_sets": mis_sets,
-        "mis_sets_labels": mis_sets_labels,
-        "mis_metrics": mis_metrics,
-        "mis_sorted": mis_sorted,
-        "mis_ranked": mis_ranked,
-        "rank_groups": rank_groups,
-        "best_mis_rank1": best_mis_rank1,
-        "best_mis_rank2": best_mis_rank2,
-        "unique_metric_values": unique_metric_values,
-        "min_component_compactness": min_compactness,
-        "component_compactness": component_metrics,
-        "homogeneity_stats": homogeneity_stats,
-        "labels": labels,
-        "alpha": alpha,
-        "N": N,
-        "M": M,
-        "sigma_z": sigma_z,
-        "z_crit": z_crit,
-        "corr_report": corr_report
-    }
 
 
-def misda_significance(Y, alpha=0.05, ensure_coverage=True, min_coverage=None):
-    """
-    Executes the Maximal Independent Structural Dimensionality Analysis (MISDA) logic.
-    """
-    if isinstance(Y, pd.DataFrame):
-        X = Y.values
-        labels = list(Y.columns)
-    else:
-        X = np.asarray(Y)
-        M = X.shape[1]
-        labels = [f"f{i+1}" for i in range(M)]
-
-    N, M = X.shape
-    corr = np.corrcoef(X, rowvar=False)
-    return misda_significance_from_corr(corr, N, M, alpha=alpha, labels=labels, ensure_coverage=ensure_coverage, min_coverage=min_coverage)
 
 
 # -------------------------------------------------------------------------
@@ -691,489 +510,7 @@ def explain_ses(out, top_k=8, name=None, show_all=False):
 # HIGH-LEVEL API
 # --------------------------------------------------------------------------------------
 
-class MISCandidate:
-    """
-    Represents a single Maximum Independent Set (MIS) solution found by the algorithm.
-    Wrapper around the internal dictionary to provide object-oriented access.
-    """
-    def __init__(self, data: dict):
-        self._data = data
 
-    @property
-    def indices(self):
-        """List of column indices corresponding to the selected variables."""
-        return self._data.get('mis_indices', [])
-
-    @property
-    def labels(self):
-        """List of variable names (column headers) of the selected variables."""
-        return self._data.get('mis_labels', [])
-
-    @property
-    def rank(self):
-        """Rank of this solution (1 = Best)."""
-        return self._data.get('rank', 999)
-
-    @property
-    def size(self):
-        """Number of variables in this solution."""
-        return len(self.indices)
-
-    @property
-    def total_correlation(self):
-        """Sum of internal pair-wise correlations (lower is better)."""
-        return self._data.get('total_correlation', float('inf'))
-
-    @property
-    def max_correlation(self):
-        """Maximum single pair-wise correlation within this set (lower is better)."""
-        return self._data.get('max_correlation', float('inf'))
-    
-    def __repr__(self):
-        return f"<MISCandidate: {self.labels} (Size={self.size}, Rank={self.rank})>"
-
-class MISDAResult:
-    """
-    Encapsulates the complete result of an MISDA analysis.
-    Stores input parameters, diagnostic regimes, execution results (MIS),
-    and validation metrics (SES).
-    """
-    def __init__(self, Y, caution, alpha_min, alpha_max, metrics, regime, alpha_exec, isda_res, name=None):
-        self.Y = Y
-        self.name = name
-        self.caution = caution
-        self.alpha_min = alpha_min
-        self.alpha_max = alpha_max
-        self.metrics = metrics
-        self.regime = regime
-        self.alpha = alpha_exec # effectively used alpha
-        self.isda_results = isda_res
-        self.validation_metrics = {}
-
-    def validate(self, check_linear=True, check_nonlinear=True, check_pareto=True):
-        """
-        Runs post-hoc validation metrics.
-        Args:
-            check_linear (bool): Run calculate_ses (Linear Regression).
-            check_nonlinear (bool): Run calculate_ses_nonlinear (Random Forest).
-            check_pareto (bool): Run evaluate_pareto_consistency.
-        """
-        # Linear SES
-        if check_linear:
-             if self.best_mis:
-                best_ids = self.best_mis.indices
-                Y_val = self.Y.values if hasattr(self.Y, "values") else self.Y
-                self.validation_metrics['linear'] = calculate_ses(Y_val, best_ids, return_details=True)
-        
-        # Non-Linear SES
-        if check_nonlinear:
-             if self.best_mis:
-                best_ids = self.best_mis.indices
-                # Safeguard: prevent massive RF runs on huge data unless explicit
-                if self.Y.shape[0] <= 10000:
-                    try:
-                        self.validation_metrics['nonlinear'] = calculate_ses_nonlinear(self.Y, best_ids, return_details=True)
-                    except ImportError:
-                        pass
-        
-        # Pareto
-        if check_pareto:
-             if self.best_mis:
-                try:
-                    p, r = evaluate_pareto_consistency(self)
-                    self.validation_metrics['pareto'] = (p, r)
-                except Exception:
-                    pass
-
-    @property
-    def ses_results(self):
-        """Backward compatibility for linear SES results."""
-        return self.validation_metrics.get('linear')
-
-    @property
-    def correlations(self):
-        """Returns the correlation report string from the ISDA execution."""
-        return self.isda_results.get('corr_report')
-
-    @property
-    def min_compactness(self):
-        """Returns the minimum component compactness found (worst internal correlation)."""
-        return self.isda_results.get('min_component_compactness', 1.0)
-    
-    @property
-    def homogeneity_ratio(self):
-        """Returns the global homogeneity ratio (worst Min/Max within a component)."""
-        stats = self.isda_results.get('homogeneity_stats', {})
-        return stats.get('min_ratio', 1.0)
-
-    @property
-    def mis_sets(self):
-        """
-        Returns a list of all MISCandidate objects found, sorted by rank.
-        """
-        raw_sets = self.isda_results.get('mis_ranked', [])
-        return [MISCandidate(d) for d in raw_sets]
-    
-    @property
-    def best_mis(self):
-        """Returns the top-ranked MISCandidate or None."""
-        if self.isda_results.get('mis_ranked'):
-            return MISCandidate(self.isda_results['mis_ranked'][0])
-        return None
-
-    @property
-    def best_mis_indices(self):
-        """Returns the list of indices of the best MIS."""
-        mis = self.best_mis
-        return mis.indices if mis else []
-
-    @property
-    def best_mis_labels(self):
-        """Returns the list of labels of the best MIS."""
-        mis = self.best_mis
-        return mis.labels if mis else []
-    
-    @property
-    def ranked_mis_sets(self):
-        """
-        Returns a dictionary mapping rank (int) -> list of MISCandidate objects.
-        """
-        raw_groups = self.isda_results.get('rank_groups', {})
-        return {r: [MISCandidate(d) for d in l] for r, l in raw_groups.items()}
-
-    def get_mis_by_rank(self, rank):
-        """
-        Returns the list of MISCandidate objects for the specified rank.
-        Returns empty list if rank not found.
-        """
-        return self.ranked_mis_sets.get(rank, [])
-        
-    @property
-    def reduction_applied(self):
-        """Boolean: True if dim(MIS) < dim(Y)."""
-        mis = self.best_mis
-        if mis:
-            return mis.size < self.Y.shape[1]
-        return False
-
-    # --- Flattened Metrics ---
-    @property
-    def separation_score(self):
-        """The Separation Score (S). Higher is better."""
-        return float(self.metrics.get("S", float('nan')))
-
-    @property
-    def normalized_separation_score(self):
-        """Normalized S-score (S_norm). Closer to 1.0 is better."""
-        return float(self.metrics.get("S_norm", float('nan')))
-
-    # --- Flattened Validation ---
-    @property
-    def ses_nonlinear_results(self):
-        """Detailed Non-Linear SES dict result. None if not run."""
-        val = self.validation_metrics.get('nonlinear')
-        return val if isinstance(val, dict) else None
-
-    @property
-    def ses_nonlinear(self):
-        """Non-Linear SES scalar metric score (float or None)."""
-        val = self.validation_metrics.get('nonlinear')
-        if isinstance(val, dict):
-            return val.get('ses')
-        return val
-    
-    @property
-    def pareto_precision(self):
-        """Pareto Precision (Safety). None if not run."""
-        p_r = self.validation_metrics.get('pareto')
-        return p_r[0] if p_r else None
-        
-    @property
-    def pareto_recall(self):
-        """Pareto Recall (Coverage). None if not run."""
-        p_r = self.validation_metrics.get('pareto')
-        return p_r[1] if p_r else None
-    
-    def to_pandas(self):
-        """
-        Exports all found independent sets to a pandas DataFrame.
-        Columns: ['rank', 'size', 'max_corr', 'total_corr', 'labels', 'indices']
-        """
-        import pandas as pd
-        data = []
-        for m in self.mis_sets:
-            data.append({
-                'rank': m.rank,
-                'size': m.size,
-                'max_corr': m.max_correlation,
-                'total_corr': m.total_correlation,
-                'labels': m.labels,
-                'indices': m.indices
-            })
-        if not data:
-            return pd.DataFrame(columns=['rank', 'size', 'max_corr', 'total_corr', 'labels', 'indices'])
-        return pd.DataFrame(data)
-
-    def __repr__(self):
-        n_start = self.Y.shape[1]
-        n_end = self.best_mis.size if self.best_mis else "?"
-        name_str = f"'{self.name}'" if self.name else "Untitled"
-        return f"<MISDAResult: {name_str} (Dim {n_start}->{n_end}, Rank={self.best_mis.rank if self.best_mis else '?'})>"
-
-    @property
-    def diagnosis(self):
-        """Returns a short diagnostic string based on Fidelity and Homogeneity."""
-        if not self.reduction_applied:
-            return "Valid (No Reduction Required)"
-
-        f = None
-        status = None
-        if self.ses_results and isinstance(self.ses_results, dict):
-            f = self.ses_results.get('F_real', None)
-            status = self.ses_results.get('status', None)
-
-        if status == "NO_REDUCTION":
-            return "Valid (No Reduction Required)"
-        
-        h = self.homogeneity_ratio
-        
-        if f is None or math.isnan(f):
-            return "Unvalidated (Missing SES)"
-
-        comps = self.isda_results.get('components_labels', [])
-        num_comps = len(comps)
-
-        # Strict clique completeness check (min_compactness >= alpha)
-        is_true_clique = (self.min_compactness >= self.alpha)
-
-        # Heuristic Decision Tree
-        if f >= 0.9 and h >= 0.8:
-            if num_comps > 1:
-                return "Ideal (Disjoint Cliques)" if is_true_clique else "Ideal (Multiple Components)"
-            return "Ideal (Clique)" if is_true_clique else "Good (Robust)"
-        if f >= 0.9 and h < 0.2:
-             return "Entangled (Mixed)"
-        if f >= 0.9:
-             return "Good (Robust)"
-             
-        if f < 0.8 and h >= 0.6:
-             return "Drift (Chain)"
-             
-        if f < 0.6 and h < 0.6:
-             return "Fragmented (Bridge)"
-             
-        return "Ambiguous/Warn"
-
-    @property
-    def validation_status(self):
-        """Returns string describing what has been validated."""
-        validated = []
-        if 'linear' in self.validation_metrics: validated.append("Linear")
-        if 'nonlinear' in self.validation_metrics: validated.append("Non-Linear")
-        if 'pareto' in self.validation_metrics: validated.append("Pareto")
-        return ", ".join(validated) if validated else "None"
-
-    def summary(self):
-        """Returns a textual summary of the analysis."""
-        lines = []
-        lines.append("\n" + "" * 70)
-        title = f"MISDA Analysis Summary: {self.name}" if self.name else "MISDA Analysis Summary"
-        lines.append(title)
-        lines.append("-" * 70)
-        
-        # Ground Truth / Inputs
-        lines.append(f"Input: [N={self.Y.shape[0]}, M={self.Y.shape[1]}]")
-        lines.append(f"Caution: {self.caution}")
-        
-        # Diagnosis
-        lines.append("\n--- 1. Diagnosis ---")
-        lines.append(describe_alpha_regime(self.metrics))
-        lines.append(f"Regime: {self.regime.name}")
-        lines.append(f"Validation: {self.validation_status}")
-        
-        # Decision
-        lines.append("\n--- 2. Decision ---")
-        if self.reduction_applied:
-            lines.append("Action: Reduction APPLIED")
-        else:
-            lines.append("Action: Full Dimension Kept (No Reduction)")
-        lines.append(f"Alpha Used: {self.alpha:.6g} (Range: [{self.alpha_min:.6g}, {self.alpha_max:.6g}])")
-                         
-        # Results
-        lines.append("\n--- 3. Results ---")
-        mis = self.best_mis
-        if mis:
-             lines.append(f"Best MIS Size: {mis.size}")
-             lines.append(f"Best MIS Labels: {mis.labels}")
-        else:
-             lines.append("No independent set found (or execution failed).")
-             
-        # Quality
-        lines.append("\n--- 4. Quality ---")
-        ratio = self.homogeneity_ratio
-        diag = self.diagnosis
-        
-        def _fmt_ratio(r):
-            if math.isnan(r): return "N/A"
-            return f"{r:.4f}"
-            
-        lines.append(f"Homogeneity Ratio: {_fmt_ratio(ratio)}")
-        lines.append(f"Auto-Diagnosis: {diag}")
-        
-        if not math.isnan(ratio) and ratio < 0.6:
-            lines.append("WARNING: Low homogeneity ratio (< 0.6). Possible over-reduction due to transitive chains or bridges.")
-        else:
-            lines.append("Status: OK (Components are internally homogeneous)")
-
-        # Global Complexity Warning (Sphere Paradox)
-        if self.reduction_applied:
-            se_norm = calculate_spectral_entropy(self.Y)
-            if se_norm > 0.75:
-                # User-requested warning
-                lines.append("WARNING: High global complexity detected (SE={:.2f}) despite aggressive reduction. Suspected Latent Conflict (Sphere-like topology).".format(se_norm))
-
-        # SES (Linear)
-        if 'linear' in self.validation_metrics:
-             lines.append("\n--- 5. Validation (SES - Linear) ---")
-             lines.append(explain_ses(self.validation_metrics['linear'], name=self.name))
-        
-        # SES (Non-Linear)
-        if 'nonlinear' in self.validation_metrics:
-             nl_res = self.validation_metrics['nonlinear']
-             if isinstance(nl_res, dict):
-                 ses_nl = nl_res.get('ses')
-                 f_real_nl = nl_res.get('F_real')
-                 nl_str = f"{ses_nl:.4f}" if ses_nl is not None else "N/A"
-                 f_str = f"{f_real_nl:.4f}" if f_real_nl is not None else "N/A"
-                 lines.append(f"Non-Linear SES (RF): {nl_str} (F_real = {f_str})")
-             else:
-                 nl_str = f"{nl_res:.4f}" if nl_res is not None else "N/A"
-                 lines.append(f"Non-Linear SES (RF): {nl_str}")
-
-        # Pareto Consistency
-        if 'pareto' in self.validation_metrics:
-             lines.append("\n--- 6. Pareto Consistency ---")
-             prec, rec = self.validation_metrics['pareto']
-             lines.append(f"Precision (Safety):   {prec:.4f}  (Prob. that Surrogate Optimum is True Optimum)")
-             lines.append(f"Recall    (Coverage): {rec:.4f}  (Prob. that True Optimum is retained)")
-             if prec < 1.0:
-                 lines.append("WARN: Surrogate introduces false optima (Precision < 1.0).")
-             if rec < 0.8:
-                 lines.append("WARN: Surrogate misses significant portion of Pareto front (Recall < 0.8).")
-         
-        return "\n".join(lines)
-
-
-    def report(self, top_k=5):
-        """
-        Returns a comprehensive technical report of the analysis.
-        Combines the standard summary with deep inspection of internal state.
-        
-        Args:
-            top_k (int): Number of candidates to show per rank (default: 5).
-        """
-        # Start with standard summary
-        base_report = self.summary()
-        
-        lines = []
-        lines.append("\n" + "=" * 70)
-        lines.append("                    DETAILED INSPECTION REPORT")
-        lines.append("=" * 70)
-        
-        # 1. Statistical Foundation
-        lines.append("\n--- A. Statistical Foundation ---")
-        lines.append(f"MISDA Version: {__version__}")
-        lines.append(f"Sample Size (N): {self.isda_results.get('N')} | Objectives (M): {self.isda_results.get('M')}")
-        lines.append(f"Fisher Transform Error (sigma_z): {self.isda_results.get('sigma_z', 0):.6f}")
-        lines.append(f"Alpha Range: [min={self.alpha_min:.3g} (Signal), max={self.alpha_max:.3g} (Noise)]")
-        lines.append(f"Caution setting: {self.caution:.2f}")
-        lines.append(f"Effective Alpha: {self.alpha:.6g} (based on regime={self.regime.name})")
-        lines.append(f"Critical Z-score: {self.isda_results.get('z_crit', 0):.4f}")
-        
-        # 2. Graph & Component Details
-        lines.append("\n--- B. Graph Topology Details ---")
-        comps = self.isda_results.get('components_labels', [])
-        homog_stats = self.isda_results.get('homogeneity_stats', {})
-        
-        lines.append(f"Connected Components: {len(comps)}")
-        for i, c in enumerate(comps):
-            # Try to get specific stats for this component if available
-            c_stat = homog_stats.get('details', {}).get(i, {})
-            min_r = c_stat.get('min_r', float('nan'))
-            max_r = c_stat.get('max_r', float('nan'))
-            ratio = c_stat.get('ratio', float('nan'))
-            
-            def _fmt_nan(v, default="N/A"):
-                if math.isnan(v): return default
-                return f"{v:.4f}"
-            
-            status = "Tight" if (not math.isnan(ratio) and ratio > 0.8) else "Loose"
-            lines.append(f"  C{i+1}: {c}")
-            lines.append(f"      Internal Correlation: [{_fmt_nan(min_r)} ... {_fmt_nan(max_r)}] | Homogeneity: {_fmt_nan(ratio)} ({status})")
-
-        # 3. Solution Space (All Candidates)
-        lines.append("\n--- C. Solution Space (All Candidates) ---")
-        rank_groups = self.ranked_mis_sets
-        
-        if not rank_groups:
-             lines.append("  No solutions found.")
-        
-        for r in sorted(rank_groups.keys()):
-            cands = rank_groups[r]
-            n_cands = len(cands)
-            lines.append(f"  Rank {r} ({n_cands} candidates):")
-            
-            # Smart Truncation
-            show_cands = cands[:top_k]
-            for c in show_cands:
-                lines.append(f"    - {c.labels} (Size={c.size})")
-                lines.append(f"      Criteria: TotalCorr={c.total_correlation:.4f} | MaxCorr={c.max_correlation:.4f}")
-            
-            if n_cands > top_k:
-                lines.append(f"      ... (+ {n_cands - top_k} more candidates. Use `res.to_pandas()` to view all.)")
-
-        # 4. Extended Verification
-        lines.append("\n--- D. Verification Details ---")
-        if self.ses_results:
-            ses = self.ses_results
-            f_real = ses.get('F_real')
-            f_null = ses.get('F_null')
-            s_val = ses.get('ses')
-            
-            def _fmt_v(val):
-                return f"{val:.4f}" if val is not None else "N/A"
-
-            lines.append("  Linear SES Breakdown:")
-            lines.append(f"    Fidelity (Real): {_fmt_v(f_real)}")
-            lines.append(f"    Fidelity (Null): {_fmt_v(f_null)}")
-            lines.append(f"    Raw SES Score:   {_fmt_v(s_val)}")
-        else:
-            lines.append("  Linear SES: Not run (or failed).")
-
-        return base_report + "\n".join(lines)
-
-    def plot(self, show=True):
-        """
-        Plots the ISDA graph.
-        
-        Args:
-            show (bool): If True, calls plt.show() to display the plot immediately.
-            
-        Returns:
-            matplotlib.figure.Figure: The figure object.
-        """
-        ret = plot_custom_misda_graph(
-            self.isda_results,
-            title=f"{self.name or 'MISDA'} — alpha={self.alpha:.3g} — regime={self.regime.name}",
-            show_removed=False
-        )
-        fig = ret['fig']
-        
-        if show:
-            import matplotlib.pyplot as plt
-            plt.show()
-            
-        return fig
 
 
 # --------------------------------------------------------------------------------------
@@ -1437,34 +774,6 @@ def _select_knee_candidate_id(frontier_candidates):
     return scored[0][5].candidate_id
 
 
-def _analyze_static_fast(Y, corr, alpha_min, alpha_max, alpha_exec, caution=1.0, name=None, ensure_coverage=True):
-    """
-    Executes static analysis using precomputed correlation matrix and alpha bounds.
-    """
-    metrics = diagnose_alpha_regime(alpha_min, alpha_max)
-    regime = AlphaRegime(metrics["regime"])
-    
-    if hasattr(Y, "values"):
-        data = np.asarray(Y.values, dtype=float)
-        labels = list(Y.columns)
-    else:
-        data = np.asarray(Y, dtype=float)
-        labels = [f"f{i+1}" for i in range(data.shape[1])]
-        
-    N, M = data.shape
-    res = misda_significance_from_corr(corr, N, M, alpha=alpha_exec, labels=labels, ensure_coverage=ensure_coverage)
-    
-    return MISDAResult(
-        Y=Y,
-        caution=caution,
-        alpha_min=alpha_min,
-        alpha_max=alpha_max,
-        metrics=metrics,
-        regime=regime,
-        alpha_exec=alpha_exec,
-        isda_res=res,
-        name=name
-    )
 
 
 def _analyze_adaptive(
@@ -1720,33 +1029,6 @@ def _analyze_adaptive(
     )
 
 
-def _analyze_static(Y, caution=1.0, name=None, ensure_coverage=True, alpha=None):
-    """
-    Executes the full MISDA pipeline on dataset Y.
-    """
-    _validate_input_matrix(Y)
-    alpha_min, alpha_max, r_max_real, r_null = estimate_alpha_interval(Y)
-    metrics = diagnose_alpha_regime(alpha_min, alpha_max)
-    regime = AlphaRegime(metrics["regime"])
-    
-    if alpha is not None:
-        alpha_exec = alpha
-    else:
-        alpha_exec = select_alpha(alpha_min, alpha_max, caution)
-    
-    res = misda_significance(Y, alpha=alpha_exec, ensure_coverage=ensure_coverage, min_coverage=None)
-    
-    return MISDAResult(
-        Y=Y,
-        caution=caution,
-        alpha_min=alpha_min,
-        alpha_max=alpha_max,
-        metrics=metrics,
-        regime=regime,
-        alpha_exec=alpha_exec,
-        isda_res=res,
-        name=name
-    )
 
 def compile_benchmark_summary(results_dict, sort_by=None):
     """
